+++ com.unity.render-pipelines.high-definition/HDRP/RenderPipeline/HDRenderPipeline.cs
@@ -447,7 +447,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                 return false;
             }
 
-            // VR is not supported currently in HD
+            // VR is not fully supported in HD
             if (XRGraphicsConfig.enabled)
             {
                 CoreUtils.DisplayUnsupportedXRMessage();
@@ -626,8 +626,11 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
                 // Set up UnityPerView CBuffer.
                 hdCamera.SetupGlobalParams(cmd, m_Time, m_LastTime, m_FrameCount);
-                if (hdCamera.frameSettings.enableStereo)
-                    hdCamera.SetupGlobalStereoParams(cmd);
+
+                // TODO VR: Current solution for compute shaders grabs matrices from
+                // stereo matrices even when not rendering stereo in order to reduce shader variants.
+                // After native fix for compute shader keywords is completed, qualify this with stereoEnabled.
+                hdCamera.SetupGlobalStereoParams(cmd);
 
                 cmd.SetGlobalVector(HDShaderIDs._IndirectLightingMultiplier, new Vector4(VolumeManager.instance.stack.GetComponent<IndirectLightingController>().indirectDiffuseIntensity, 0, 0, 0));
 
@@ -1111,8 +1114,11 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                             // Overwrite camera properties set during the shadow pass with the original camera properties.
                             renderContext.SetupCameraProperties(camera, hdCamera.frameSettings.enableStereo);
                             hdCamera.SetupGlobalParams(cmd, m_Time, m_LastTime, m_FrameCount);
-                            if (hdCamera.frameSettings.enableStereo)
-                                hdCamera.SetupGlobalStereoParams(cmd);
+
+                            // TODO VR: Current solution for compute shaders grabs matrices from
+                            // stereo matrices even when not rendering stereo in order to reduce shader variants.
+                            // After native fix for compute shader keywords is completed, qualify this with stereoEnabled.
+                            hdCamera.SetupGlobalStereoParams(cmd);
                         }
 
                         using (new ProfilingSample(cmd, "Screen space shadows", CustomSamplerId.ScreenSpaceShadows.GetSampler()))